Added punch sprites
for reference, the gifs i sent used the following timing in each frame. but feel free to adjust, especially since the wind-up frames will be determined by how long the player holds the button.
1: 400 ms
2: 200 ms
3: 200 ms
4: 300 ms
5: 200 ms
6: 200 ms
7: 300 ms
Also, the first frame is just the standing still frame of each direction.
Also also, there's duplicates on this sprite sheet just to make it easier to know the order of things.